vm_control: flush IRQs to LAPIC for snapshot.
Just prior to taking a snapshot, we want to ensure that all pending IRQs
are flushed to LAPICs so that they can be safely saved & restored with a
snapshot. This CL wires the snapshot handler to the IRQ handler so that
the IRQs can be confirmed to be flushed.

BUG=b:269652620
TEST=./tools/dev_container e2e_tests/run --tests suspend_resume exercises this
code path, but does not directly test the changed behavior.
TEST=this fixes a race condition. Writing a test to reproduce that race
condition would just introduce a flakey test. When we have a cold boot
test (which we can't write yet because it depends future features), it
will flake on restoring LAPIC state *if* this CL has not fixed the race
condition properly. Thus, the testing strategy for this CL is to make
sure our cold boot test is run with *many* iterations to detect such
flakes (see b/270778349).

Change-Id: I2f326a7fdee9f21edda8402920661ef984943761
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4292623
Reviewed-by: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>