devices: block: fix race condition causing panic on shutdown
This should fix an error we've been seeing in Android:

    thread 'virtio_blk' panicked at 'too many refs to the disk'

When `Task::detach` is used, we can't manage the lifetime of detached
task and when used in combination with `BlockingPool`, it seems to be
possible to leak the `BlockingPool` or even the whole `Executor`. In
particular, when shutting down crosvm (or, in the future, putting the
block device to sleep), if there are active detached tasks, we end up
not being able to reclaim the `Rc<DiskState>` because the leaked tasks
can still contain clones of it.

BUG=b:270228782

Change-Id: I6d4a88eb375017de03b9c60be7130abdbf1d0038
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4292624
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>