vm_control: snapshot: fix tube response deserialization
The tube response should be deserialized as a `SnapshotControlResult`
instead of `()`.

Also, we want to flush IRQs before snapshotting the IRQ chip state, so
it should be done before sending the snapshot command. However, this
part is still not right because the devices aren't asleep.

Change-Id: I94245b23a5166ec231dd36020e55930a80e007cd
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4304682
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>