devices: virtio-net: use features in queue count calculation
The virtio specification for the network device allows the driver to
negotiate two features related to queue counts:
- VIRTIO_NET_F_MQ enables use of more than one rx/tx pair
- VIRTIO_NET_F_CTRL_VQ enables an additional control queue

The crosvm implementation of virtio-net assumed that the driver would
support both of these features and neglected to check the result of the
feature negotiation process; this would result in the net device failing
to start if a driver did not enable the CTRL_VQ feature and did not
provide the extra queue required when that feature is enabled.
Additionally, if the driver did not negotiate the MQ feature, the device
would fail to start up if the maximum supported number of queue pairs
was greater than one.

Fix both of these issues by checking the features acknowledged by the
driver at activate() time and adjusting the expected number of queues
accordingly. As part of this change, the control queue becomes optional,
and the previous code used the presence of the control queue on a worker
thread to indicate whether to perform interrupt resampling work; this
change splits that into a separate handle_interrupt_resample flag passed
to the worker run() function so that the interrupt is resampled even
when the CTRL_VQ flag is not negotiated.

BUG=b:272536271
BUG=b:164505993
TEST=tools/dev_container tools/presubmit

Change-Id: I3dd65c7b134c75f5fc6881ed33c7b8df83c48c36
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4326887
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dmitrii Merkurev <dimorinny@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>