devices: vvu proxy: move worker thread out of State
The worker thread handle was part of the State object, and when the
worker thread was exiting, it would change the state from State::Running
(which contained the WorkerThread) to another state, causing that
WorkerThread instance to be dropped. The Drop handler of WorkerThread
joins the thread, which is illegal when called from within the thread
itself, so this caused an EDEADLK error and a panic.

Move the WorkerThread to the top-level VirtioVhostUser structure, which
is only accessible from the main thread, to avoid this problem.

BUG=b:267980534
TEST=Shut down Crostini on brya-manatee

Change-Id: I7dfca3e0a92c97bbd5aad2c48da191b0b1b30461
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4330489
Reviewed-by: Tomasz Nowicki <tnowicki@google.com>
Commit-Queue: Tomasz Nowicki <tnowicki@google.com>