devices: snapshot serializion for MsixConfig.
MsixConfig is now serializable. After deserialization, there is a new function
`restore_msix_all`, which will re-create the MSI-X vectors based on the
data that was deserialized. We could try to weave this into the actual
serde deserialization process, but that will likely be more hassle than
it is worth (it will also stop us from fixing up the msi control Tube on
restore, which we'll have to do in a future CL).

BUG=b:272137957
TEST=presubmit

Change-Id: I36fbc35a13fe11ae2f7c899431d2ef239ac00f90
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4337964
Reviewed-by: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>