crosvm: improve cold restore
This change tweaks the `--restore` flag to be compatible with the
(upcoming) vCPU restore code. Restoring the vCPUs requires them to have
been started (otherwise we'll essentially deadlock), so now we do the
following:

* Tell the vCPUs to start in a suspended state.
* Wait for vCPU init to finish.
* Perform the restore.
* Tell the vCPUs to start running.

I've left the Windows half unimplemented because there are some missing
pieces + I expect a lot of churn in this area and so it will probably be
less work overall to port the restore code once the Linux half is
actually working.

BUG=b:269174485

Change-Id: I07588e887789498a951d9427391415afcb646e96
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4343614
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>
Reviewed-by: Elie Kheirallah <khei@google.com>