virtio: block: fix `id` field serialization/deserialization
Since this field is an option, it can be specified as a `null` value in
JSON. Trying to deserialize this directly as a String will not play
well, so make sure to specify the correct type for deserialization.

Adding unit tests also revealed that the serialization function was
missing (otherwise `id` gets serialized into a sequence of bytes, which
the deserializer cannot recognize), so add it.

BUG=b:274376388
TEST=crosvm can boot a VM without triggering a panic when
crrev.com/c/4349368 is applied.

Change-Id: Iaa4c99fc23375afd8c93e68573330deef7c8fe51
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4349418
Reviewed-by: Ningyuan Wang <ningyuan@google.com>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>