devices: pci_root: fix bit shifts in IO config word write
The PciConfigIo implementation of PCI configuration mechanism #1 (the
traditional port I/O method using port 0xCF8 and 0xCFC) previously tried
to account for byte- and word-sized writes to the configuration address
register (0xCF8) in addition to the expected dword-sized writes required
by the spec. The word-sized write path was treating the offset provided
to the set_config_address() function as a word offset (multiples of 16
bits), but it was in fact always a byte offset (multiples of 8 bits);
this would result in a runtime panic if the guest performed a word-size
write to the upper 16 bits of the 0xCF8 register (offset = 2).

Fix the shift count so that it is multiplying offset by 8 in all cases
so that the shift will never overflow (taking into account the offset +
len check at the top of the function).

BUG=b:274366589
TEST=tools/dev_container tools/presubmit

Change-Id: Ie124b2a339b03bfba2eaae292263021b36b14972
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4356875
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>