devices: bus: avoid potential add overflow in BusRange
Avoid potential overflow when adding the base and length of a range by
using saturating addition.

If the total of range base and length exceeds the size of a u64, this
usually indicates an incorrect range, but using saturating adds allows
us to handle this gracefully rather than causing a panic when checked
arithmetic is enabled.

BUG=b:274366589
TEST=tools/dev_container tools/presubmit

Change-Id: I358d906718d3e8c2d7a5c72721b095199fc785cb
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4356903
Reviewed-by: Zihan Chen <zihanchen@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>