devices: save/restore MsixConfig w/ VirtioPciDevice.
Now that we have a way to recreate MsixConfig from a snapshot, this CL
wires up the save/restore code to include it in the snapshot of a
VirtioPciDevice (previously we only had the inner device). We've also
refactored how snapshot/restore works so they're more consistent with
how the rest of the snapshot subsystem expects to use them.

BUG=b:272139104
TEST=presubmit

Change-Id: I7c3f666c626aac6c7090f0bf62c10249f19e788e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4355968
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Elie Kheirallah <khei@google.com>