crosvm: Make cpuset cgroup path conditional
The preceding patch on:
https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4304743

made the assumption that the cpuset cgroup path would always be
available. This patch will correct this assumption and only
add it into the sandbox if the path exists.

This particular path is not necessary if one does not wish
to make use of the gpu cgroup command line options.

BUG=b:269139377
TEST=Launch Borealis VM and ensured gpu cgroup functionality works
as expected.

Change-Id: I52de208b0c2b860c696955297b056aed0a0edf1e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4357919
Reviewed-by: Frederick Mayle <fmayle@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>