cmdline: unify hypervisor options in --hypervisor
Enhance the --hypervisor option so it can also specify the device path
for the particular hypervisor (e.g. --hypervisor=kvm[device=/dev/kvm]).

This replaces the functionality of the existing --kvm-device option,
which is now deprecated, and the newly-added --geniezone-device option,
which is removed since there should be no existing users.

Existing users of the --hypervisor option do not need to make any
changes, as the extended option enum is compatible with the previous
command line option.

BUG=b:274817652
TEST=tools/dev_container tools/presubmit

Change-Id: Iefb855bc136632526daf455ae033b63fcc6fbe59
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4363324
Reviewed-by: Noah Gold <nkgold@google.com>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>