devices: MsixConfig: clear vectors on restore.
The initial restore code just discarded any existing vectors on restore.
This is fine in the case of a cold restore, but in the case of a warm
restore is not a good idea. It could leak IRQ lines, and/or fail to
create the desired GSI because it was previously registered with
different routing data. To fix this, we explicitly release all vectors
on restore.

BUG=b:276355925
TEST=presubmit

Change-Id: Id8e671e905e9ef1a86f31db582abba8246c2036d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4385819
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>