jail: avoid extra string conversions in build.rs
Assign the env() result to a variable so it lives long enough that we
can use a &str reference into it rather than using .to_owned() again.
No need to convert to &str and then back to String for the cases that
don't need to be modified.

Also fixes a new clippy warning about an unnecessary reference in the
PathBuf::join() call.

BUG=b:276487055
TEST=tools/clippy # with rust 1.68

Change-Id: I385bdd4dc0cde41937195630f3d5d9d93ad00468
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4390737
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Zihan Chen <zihanchen@google.com>