crosvm: gpu: simplify wayland socket bind mounts
The previous code made a copy of the list of socket paths using
collect(), but we still have the original BTreeMap containing the socket
paths in cfg, so we can just iterate over the values in place.

Additionally, if a socket path does not have a valid parent, produce an
error message that includes the invalid path to help with debugging.

BUG=None
TEST=tools/presubmit

Change-Id: If799c2298caf22eccf0648c6b66fc7b201cea63b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4390734
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Zihan Chen <zihanchen@google.com>