devices: virtio: Serialize PassthroughFs unit tests with NamedLock
Use NamedLock to prevent PassthroughFs tests from running in
multi-threads when one runs `cargo test`.
Previously, we ran these tests as a part of integration tests with
`libtest_mimic` to ensure it's not parallelized. But, using nextest's
feature is more straightforward. Also, running them as unit tests allow
us to call mocked functions with `#[cfg(test)]` in the future.

Note that NamedLock is not necessary for `cargo nextest` because it runs
each test single-threaded, but in multi-processes.

BUG=b:275016855
BUG=b:277302336
TEST=run_tests

Change-Id: I3751b712eaf1b2180bf4153ab501a32b402c0a29
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4406098
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>