devices: virtio: create Reader and Writer in DescriptorChain
This is a follow-up to commit 63f50362ec21 ("devices: virtio:
overhaul DescriptorChain API"). With this change, the normal case of
creating a single Reader and Writer per DescriptorChain, there is no
need to copy the MemRegion lists, which minimizes the extra allocation
and copying overhead.

Additionally, the ExportedRegion list is now stored in an Arc, so no
copying is required; instead, the reference count is incremented.

BUG=None
TEST=tools/dev_container tools/presubmit

Change-Id: Ibd0b2bfc5bfd2d8c84e7ed4d024e314cbdf33a81
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4510437
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>