devices: virtio: add notification region fallback in write_bar()
Normally, writes to the virtqueue notification region should be handled
via ioevents and not make it into the write_bar() handler. However, in
an upcoming change, ioevent registration will be deferred until
activate(), and in this case, it is possible that a queue is notified
while the ioevent registration is still in progress. To handle this
case, add a fallback to the VirtioPciDevice::write_bar() function to
signal the corresponding queue event. (These processes can overlap
because the BAR writes used to configure the virtio-pci device are
posted, meaning the guest does not need to wait after writing the
configuration register to activate the device and writing to a queue
notification region to start using a queue.)

BUG=b:278130202
TEST=crosvm run bzImage # observe working virtio devices

Change-Id: Ic1b19ac54f6dace29705b9d190d280eb5e962b5e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4422465
Reviewed-by: Ningyuan Wang <ningyuan@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>